% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_plots.R
\name{get_tier_name}
\alias{get_tier_name}
\title{Convert generic tier labels to more descriptive labels.}
\usage{
get_tier_name(labels)
}
\arguments{
\item{labels}{A default, nondescriptive label such as "X1" meaning usage in tier 1,
or "B1" meaning bill amount from tier 1.}
}
\value{
A string representing the name of a tier, "Tier 1", "Tier 3", etc.
}
\description{
\code{get_tier_name} takes a generic label, like "X1" or "B1", that is
generated when caculating tiered charges, and converts it to a more
descritive label for use when plotting.
}

