% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_barchart_by_tiers}
\alias{plot_barchart_by_tiers}
\title{Barchart showing revenue/usage by tier.}
\usage{
plot_barchart_by_tiers(data, display_type, bar_type)
}
\arguments{
\item{data}{The dataframe filtered by date range and rate code.}

\item{display_type}{A string specifying whether to display "Revenue" in millions
of dollars, of "Usage" in thousands of acre-feet.}

\item{bar_type}{A string showing whether to display the "Absolute" revenue
in each tier, or the "Percent" of revenue in each tier.}
}
\value{
A plotly object created from a ggplot chart, with plotly's
modebar removed.
}
\description{
\code{plot_barchart_by_tiers} returns a bar chart showing the revenue/usage
generated under the baseline and hypthetical rate structures. These bars are
further broken down and colored by tier, so that the amount of revenue/usage
generated in each tier is visible.
}

