% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_bill_change_boxplot}
\alias{plot_bill_change_boxplot}
\title{Boxplot of bill changes.}
\usage{
plot_bill_change_boxplot(data)
}
\arguments{
\item{data}{The dataframe filtered by date range and rate code.}
}
\value{
A plotly object created from a ggplot chart, with plotly's
modebar removed.
}
\description{
\code{plot_bill_change_histogram} returns a small boxplot of changes
(hypothetical - baseline) in total amount paid during the time period
for each customer. Designed to complement \code{\link{plot_bill_change_histogram}}.
}

