% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_revenue_over_time}
\alias{plot_revenue_over_time}
\title{Line chart of revenue over time.}
\usage{
plot_revenue_over_time(data)
}
\arguments{
\item{data}{The dataframe filtered by date range and rate code.}
}
\value{
A plotly object created from a ggplot chart, with plotly's
modebar removed.
}
\description{
\code{plot_revenue_over_time} returns a line chart showing revenue over
time for both baseline and hypothetical rate structures. The baseline
rate is shown in black and the hypthotical in blue.
}

